requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pyjwt==2.8.0
bcrypt==4.1.2
pydantic==2.6.1
python-multipart==0.0.9
pytest
httpx
websockets
psycopg
langchain-text-splitters 
langchain-community 
langgraph
langchain-openai
psycopg-binary
pgvector
pdfplumber
python-dotenv
psycopg_pool